Fix linux issues, when importing pyautogui would cause errors
